Build UI and Load data for ActorList
